+++ packages/sampler/stories/moonstone-stories/DatePicker.js
@@ -1,11 +1,14 @@
-import DatePicker from '@enact/moonstone/DatePicker';
+import DatePicker, {DatePickerBase} from '@enact/moonstone/DatePicker';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
-import {boolean, text} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
-import nullify from '../../src/utils/nullify.js';
+import {boolean, text} from '../../src/enact-knobs';
+import {mergeComponentMetadata, removeProps} from '../../src/utils';
+
+const Config = mergeComponentMetadata('DatePicker', DatePickerBase, DatePicker);
+removeProps(Config, 'year defaultOpen day maxDays maxMonths month onChangeDate onChangeMonth onChangeYear order');
 
 DatePicker.displayName = 'DatePicker';
 
@@ -17,18 +20,18 @@ storiesOf('Moonstone', module)
 			text: 'The basic DatePicker'
 		})(() => (
 			<DatePicker
-				dayAriaLabel={nullify(text('dayAriaLabel', ''))}
-				dayLabel={nullify(text('dayLabel', ''))}
-				monthAriaLabel={nullify(text('monthAriaLabel', ''))}
-				monthLabel={nullify(text('monthLabel', ''))}
-				noLabels={nullify(boolean('noLabels', false))}
-				noneText={text('noneText', 'Nothing Selected')}
+				dayAriaLabel={text('dayAriaLabel', Config)}
+				dayLabel={text('dayLabel', Config)}
+				monthAriaLabel={text('monthAriaLabel', Config)}
+				monthLabel={text('monthLabel', Config)}
+				noLabels={boolean('noLabels', Config)}
+				noneText={text('noneText', Config, 'Nothing Selected')}
 				onChange={action('onChange')}
 				onClose={action('onClose')}
 				onOpen={action('onOpen')}
-				title={text('title', 'Date')}
-				yearAriaLabel={nullify(text('yearAriaLabel', ''))}
-				yearLabel={nullify(text('yearLabel', ''))}
+				title={text('title', Config, 'Date')}
+				yearAriaLabel={text('yearAriaLabel', Config)}
+				yearLabel={text('yearLabel', Config)}
 			/>
 		))
 	);