+++ packages/sampler/stories/moonstone-stories/DayPicker.js
@@ -2,10 +2,9 @@ import DayPicker from '@enact/moonstone/DayPicker';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
-import {boolean, select, text} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
-import nullify from '../../src/utils/nullify.js';
+import {boolean, select, text} from '../../src/enact-knobs';
 
 DayPicker.displayName = 'DayPicker';
 
@@ -17,17 +16,17 @@ storiesOf('Moonstone', module)
 			text: 'The basic DayPicker'
 		})(() => (
 			<DayPicker
-				aria-label={nullify(text('aria-label', ''))}
+				aria-label={text('aria-label', DayPicker)}
 				dayNameLength={select('dayNameLength', ['short', 'medium', 'long', 'full'], 'long')}
-				disabled={boolean('disabled', false)}
-				everyDayText={nullify(text('everyDayText', ''))}
-				everyWeekdayText={nullify(text('everyWeekdayText', ''))}
-				everyWeekendText={nullify(text('everyWeekendText', ''))}
-				noneText={text('noneText', 'none')}
+				disabled={boolean('disabled', DayPicker)}
+				everyDayText={text('everyDayText', DayPicker)}
+				everyWeekdayText={text('everyWeekdayText', DayPicker)}
+				everyWeekendText={text('everyWeekendText', DayPicker)}
+				noneText={text('noneText', DayPicker, 'none')}
 				onClose={action('onClose')}
 				onOpen={action('onOpen')}
 				onSelect={action('onSelect')}
-				title={text('title', 'Day Picker')}
+				title={text('title', DayPicker, 'Day Picker')}
 			/>
 		))
 	);