+++ packages/sampler/stories/moonstone-stories/Dialog.js
@@ -1,12 +1,16 @@
 import Dialog from '@enact/moonstone/Dialog';
+import Popup from '@enact/moonstone/Popup';
 import BodyText from '@enact/moonstone/BodyText';
 import Button from '@enact/moonstone/Button';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
-import {boolean, select, text} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
+import {boolean, select, text} from '../../src/enact-knobs';
+import {mergeComponentMetadata} from '../../src/utils';
+
+const Config = mergeComponentMetadata('Dialog', Popup, Dialog);
 Dialog.displayName = 'Dialog';
 
 storiesOf('Moonstone', module)
@@ -18,18 +22,22 @@ storiesOf('Moonstone', module)
 		})(() => (
 			<div>
 				<Dialog
-					casing={select('casing', ['preserve', 'sentence', 'word', 'upper'], 'upper')}
-					noAnimation={boolean('noAnimation', false)}
-					noAutoDismiss={boolean('noAutoDismiss', false)}
-					noDivider={boolean('noDivider', false)}
+					casing={select('casing', ['preserve', 'sentence', 'word', 'upper'], Config, 'upper')}
+					// null issue
+					noAnimation={boolean('noAnimation', Config)}
+					// null issue
+					noAutoDismiss={boolean('noAutoDismiss', Config)}
+					// null issue
+					noDivider={boolean('noDivider', Config)}
 					onClose={action('onClose')}
-					open={boolean('open', false)}
-					showCloseButton={boolean('showCloseButton', false)}
+					// null issue
+					open={boolean('open', Config)}
+					showCloseButton={boolean('showCloseButton', Config)}
 				>
-					<title>{text('title', 'Hello Dialog')}</title>
-					<titleBelow>{text('titleBelow', 'This is an organized dialog')}</titleBelow>
+					<title>{text('title', Config, 'Hello Dialog')}</title>
+					<titleBelow>{text('titleBelow', Config, 'This is an organized dialog')}</titleBelow>
 					<span>This dialog has content in it and can be very useful for organizing information
-					for the user.</span>
+							for the user.</span>
 					<buttons>
 						<Button>Ok</Button>
 						<Button>Nevermind</Button>