+++ packages/sampler/stories/moonstone-stories/EditableIntegerPicker.js
@@ -1,13 +1,14 @@
-import EditableIntegerPicker from '@enact/moonstone/EditableIntegerPicker';
+import EditableIntegerPicker, {EditableIntegerPickerBase} from '@enact/moonstone/EditableIntegerPicker';
 import {decrementIcons, incrementIcons} from './icons';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
-import {boolean, number, text, select} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
-import nullify from '../../src/utils/nullify.js';
+import {boolean, number, select, text} from '../../src/enact-knobs';
+import {mergeComponentMetadata} from '../../src/utils';
 
+const Config = mergeComponentMetadata('EditableIntegerPicker', EditableIntegerPickerBase, EditableIntegerPicker);
 EditableIntegerPicker.displayName = 'EditableIntegerPicker';
 
 // Set up some defaults for info and knobs
@@ -25,21 +26,21 @@ storiesOf('Moonstone', module)
 			text: 'Basic usage of EditableIntegerPicker'
 		})(() => (
 			<EditableIntegerPicker
-				decrementIcon={nullify(select('decrementIcon', ['', ...decrementIcons]))}
+				decrementIcon={select('decrementIcon', ['', ...decrementIcons], Config)}
 				defaultValue={20}
-				disabled={boolean('disabled', false)}
-				incrementIcon={nullify(select('incrementIcon', ['', ...incrementIcons]))}
-				max={number('max', 100)}
-				min={number('min', 0)}
-				noAnimation={nullify(boolean('noAnimation', false))}
+				disabled={boolean('disabled', Config)}
+				incrementIcon={select('incrementIcon', ['', ...incrementIcons], Config)}
+				max={number('max', Config, 100)}
+				min={number('min', Config, 0)}
+				noAnimation={boolean('noAnimation', Config)}
 				onBlur={action('onBlur')}
 				onChange={action('onChange')}
 				onKeyDown={action('onKeyDown')}
-				orientation={select('orientation', prop.orientation, 'horizontal')}
+				orientation={select('orientation', prop.orientation, Config)}
 				step={number('step', 1)}
-				unit={nullify(text('unit', ''))}
-				width={nullify(select('width', prop.width,  prop.width[2]))}
-				wrap={nullify(boolean('wrap', false))}
+				unit={text('unit', '')}
+				width={select('width', prop.width,  Config)}
+				wrap={boolean('wrap', Config)}
 			/>
 		))
 	);